improve the delete feature

+++ src/components/Blog.jsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import blogs from '../services/blogs';
 import Notification from './Notification';
 
-export default function Blog({ blog }) {
+export default function Blog({ blog, setBlogs }) {
   const [showBlog, setShowBlog] = useState(false);
   const [localBlog, setLocalBlog] = useState(blog);
   const handleShow = () => {
@@ -25,10 +25,20 @@ export default function Blog({ blog }) {
 
   const handleDelete = async () => {
     try {
-      const response = await blogs.deleteBlog(blog.id);
-      if (response.status === 204) {
-        Notification.success('Blog deleted successfully');
+      // confirm the deletion
+      const isConfirmed = await Notification.delete(
+        `Are you sure you want to delete this blog?`
+      );
+      if (!isConfirmed) {
+        return;
       }
+      // Call the deleteBlog function
+      const response = await blogs.deleteBlog(blog.id);
+      setBlogs((prevBlogs) => prevBlogs.filter((b) => b.id !== blog.id));
+
+      // Check the response status and show a success message
+      response.status === 204 &&
+        Notification.success('Blog deleted successfully');
     } catch (error) {
       Notification.error(error.response?.data?.error || 'An error occurred');
     }
@@ -37,22 +47,26 @@ export default function Blog({ blog }) {
   return (
     <div className="p-4 blog">
       <div className="p-6 max-w-lg mx-auto bg-white shadow-md rounded-lg">
-        <div className="flex justify-between items-center mb-4">
-          <h2 className="text-xl font-semibold">{blog.title}</h2>
-          <button
-            onClick={handleShow}
-            className={`px-3 py-1 rounded transition duration-200 ${
-              showBlog ? 'bg-gray-500 text-white' : 'bg-green-700 text-white'
-            }`}
-          >
-            {showBlog ? 'Hide' : 'Show'}
-          </button>
+        <div className="mb-4">
+          <div className="flex justify-between items-center">
+            <h2 className="text-xl font-semibold">{blog.title}</h2>
+            <button
+              onClick={handleShow}
+              className={`px-3 py-1 rounded transition duration-200 ${
+                showBlog ? 'bg-gray-500 text-white' : 'bg-green-700 text-white'
+              }`}
+            >
+              {showBlog ? 'Hide' : 'Show'}
+            </button>
+          </div>
+          <div className="mt-2">
+            <span className="text-lg font-light text-gray-700">
+              Author: {blog.author}
+            </span>
+          </div>
         </div>
 
         <div className={showBlog ? 'visible' : 'hidden'}>
-          <p className="text-lg text-gray-700 mb-2">
-            Author: <span className="font-medium">{blog.author}</span>
-          </p>
           <p className="text-lg text-gray-700 mb-2">
             Username:{' '}
             <span className="font-medium">{blog?.user?.username}</span>
